Floyd Warshall Method

+++ Phitron Algo/Module 10/Floyd_Warshall.cpp
@@ -15,7 +15,7 @@ int main()
                 adj_mat[i][j] = 0;
             }
             else
-                adj_mat[i][j] = INT_MAX;
+                adj_mat[i][j] = LONG_MAX;
         }
     }
 
@@ -33,7 +33,7 @@ int main()
         {
             for (int j = 0; j < n; j++)
             {
-                if (adj_mat[i][k] != INT_MAX && adj_mat[k][j] != INT_MAX && adj_mat[i][k] + adj_mat[k][j] < adj_mat[i][j])
+                if (adj_mat[i][k] != LONG_MAX && adj_mat[k][j] != LONG_MAX && adj_mat[i][k] + adj_mat[k][j] < adj_mat[i][j])
                 {
                     adj_mat[i][j] = adj_mat[i][k] + adj_mat[k][j];
                 }
@@ -45,7 +45,7 @@ int main()
     {
         for (int j = 0; j < n; j++)
         {
-            if (adj_mat[i][j] == INT_MAX)
+            if (adj_mat[i][j] == LONG_MAX)
             {
                 cout << "InF ";
             }
