+++ Phitron Algo/Module 10/shortest_route_2.cpp
@@ -15,7 +15,7 @@ int main()
                 adj_mat[i][j] = 0;
             }
             else
-                adj_mat[i][j] = INT_MAX;
+                adj_mat[i][j] = LONG_MAX;
         }
     }
 
@@ -33,7 +33,7 @@ int main()
         {
             for (int j = 0; j < n; j++)
             {
-                if (adj_mat[i][k] != INT_MAX && adj_mat[k][j] != INT_MAX && adj_mat[i][k] + adj_mat[k][j] < adj_mat[i][j])
+                if (adj_mat[i][k] != LONG_MAX && adj_mat[k][j] != LONG_MAX && adj_mat[i][k] + adj_mat[k][j] < adj_mat[i][j])
                 {
                     adj_mat[i][j] = adj_mat[i][k] + adj_mat[k][j];
                 }
@@ -45,7 +45,7 @@ int main()
     {
         int s,d;
         cin >> s >> d;
-        if (adj_mat[s][d] == INT_MAX)
+        if (adj_mat[s][d] == LONG_MAX)
         {
             cout << -1 << endl;
         }
